+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 Changed
 ~~~~~~~
 
+- Makes the ``pyroomacoustics.utilities.resample`` backend is made configurable
+  to avoid ``soxr`` dependency. The resample backend is configurable to
+  ``soxr``, ``samplerate``, if these packages are available, and otherwise
+  falls back to ``scipy.signal.resample_poly``.
 - Fixes typo in ``materials.json``: ``ceiling_fibre_abosrber -> ceiling_fibre_absorber``.
 
 `0.8.1`_ - 2024-10-30